refactor(core): extract pure Rust logic into core-lib crate
Split crates/core into two crates:
- crates/core-lib: pure Rust compression logic (no FFI dependencies)
- crates/core: napi-rs thin wrappers delegating to core-lib

This prepares for adding a wasm-bindgen crate that shares the same
core logic, enabling browser usage without SharedArrayBuffer.

Part of #319
