+++ crates/core-lib/src/detect.rs
@@ -1,0 +1,161 @@
+//! Unified auto-detect decompression API.
+//!
+//! Detects the compression format from magic bytes and decompresses accordingly.
+
+use crate::ComprsError;
+
+/// Zstd magic number: 0xFD2FB528 (little-endian).
+const ZSTD_MAGIC: [u8; 4] = [0x28, 0xB5, 0x2F, 0xFD];
+
+/// Gzip magic number: 0x1F 0x8B.
+const GZIP_MAGIC: [u8; 2] = [0x1F, 0x8B];
+
+/// LZ4 frame magic number: 0x184D2204.
+const LZ4_MAGIC: [u8; 4] = [0x04, 0x22, 0x4D, 0x18];
+
+/// Compression format detected from input data.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Format {
+    Zstd,
+    Gzip,
+    Brotli,
+    Lz4,
+    Unknown,
+}
+
+impl std::fmt::Display for Format {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Format::Zstd => write!(f, "zstd"),
+            Format::Gzip => write!(f, "gzip"),
+            Format::Brotli => write!(f, "brotli"),
+            Format::Lz4 => write!(f, "lz4"),
+            Format::Unknown => write!(f, "unknown"),
+        }
+    }
+}
+
+/// Detect the compression format of the given data.
+pub fn detect(data: &[u8]) -> Format {
+    if data.len() >= 4 && data[..4] == ZSTD_MAGIC {
+        return Format::Zstd;
+    }
+    if data.len() >= 2 && data[..2] == GZIP_MAGIC {
+        return Format::Gzip;
+    }
+    if data.len() >= 4 && data[..4] == LZ4_MAGIC {
+        return Format::Lz4;
+    }
+    // Brotli has no magic bytes. Attempt heuristic detection:
+    if !data.is_empty() && is_likely_brotli(data) {
+        return Format::Brotli;
+    }
+    Format::Unknown
+}
+
+/// Heuristic check for brotli data.
+/// Attempts to decompress a small prefix to see if brotli accepts the header.
+fn is_likely_brotli(data: &[u8]) -> bool {
+    use std::io::Read;
+    let mut decompressor = brotli::Decompressor::new(data, 4096);
+    let mut buf = [0u8; 1];
+    decompressor.read_exact(&mut buf).is_ok()
+}
+
+/// Decompress data by auto-detecting the compression format.
+pub fn decompress(data: &[u8]) -> Result<Vec<u8>, ComprsError> {
+    match detect(data) {
+        Format::Zstd => crate::zstd::decompress(data),
+        Format::Gzip => crate::gzip::decompress(data),
+        Format::Brotli => crate::brotli::decompress(data),
+        Format::Lz4 => crate::lz4::decompress(data),
+        Format::Unknown => Err(ComprsError::InvalidArg(
+            "unable to detect compression format; use algorithm-specific functions (zstdDecompress, gzipDecompress, brotliDecompress, lz4Decompress) instead".to_string(),
+        )),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::io::Write;
+
+    #[test]
+    fn detect_zstd() {
+        let data = zstd::bulk::compress(b"test data for zstd", 3).unwrap();
+        assert_eq!(detect(&data), Format::Zstd);
+    }
+
+    #[test]
+    fn detect_gzip() {
+        let mut encoder = flate2::write::GzEncoder::new(Vec::new(), flate2::Compression::default());
+        encoder.write_all(b"test data for gzip").unwrap();
+        let data = encoder.finish().unwrap();
+        assert_eq!(detect(&data), Format::Gzip);
+    }
+
+    #[test]
+    fn detect_brotli() {
+        let mut output = Vec::new();
+        {
+            let mut compressor = brotli::CompressorWriter::new(&mut output, 4096, 6, 22);
+            compressor.write_all(b"test data for brotli").unwrap();
+        }
+        assert_eq!(detect(&output), Format::Brotli);
+    }
+
+    #[test]
+    fn detect_lz4() {
+        let mut compressed = Vec::new();
+        let mut encoder = lz4_flex::frame::FrameEncoder::new(&mut compressed);
+        encoder.write_all(b"test data for lz4").unwrap();
+        encoder.finish().unwrap();
+        assert_eq!(detect(&compressed), Format::Lz4);
+    }
+
+    #[test]
+    fn detect_unknown() {
+        let data = b"this is not compressed data at all";
+        assert_eq!(detect(data), Format::Unknown);
+    }
+
+    #[test]
+    fn detect_empty() {
+        assert_eq!(detect(b""), Format::Unknown);
+    }
+
+    #[test]
+    fn detect_too_short() {
+        assert_eq!(detect(&[0x00]), Format::Unknown);
+    }
+
+    #[test]
+    fn decompress_auto_detect() {
+        let original = b"Hello auto-detect!";
+
+        // zstd
+        let compressed = zstd::bulk::compress(original, 3).unwrap();
+        let decompressed = decompress(&compressed).unwrap();
+        assert_eq!(original.as_slice(), decompressed.as_slice());
+
+        // gzip
+        let compressed = crate::gzip::compress(original, None).unwrap();
+        let decompressed = decompress(&compressed).unwrap();
+        assert_eq!(original.as_slice(), decompressed.as_slice());
+
+        // brotli
+        let compressed = crate::brotli::compress(original, None).unwrap();
+        let decompressed = decompress(&compressed).unwrap();
+        assert_eq!(original.as_slice(), decompressed.as_slice());
+
+        // lz4
+        let compressed = crate::lz4::compress(original).unwrap();
+        let decompressed = decompress(&compressed).unwrap();
+        assert_eq!(original.as_slice(), decompressed.as_slice());
+    }
+
+    #[test]
+    fn decompress_unknown_format() {
+        assert!(decompress(b"not compressed").is_err());
+    }
+}
