+++ crates/core/src/detect.rs
@@ -6,16 +6,7 @@ use napi::Task;
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
 
-use crate::ComprsError;
-
-/// Zstd magic number: 0xFD2FB528 (little-endian).
-const ZSTD_MAGIC: [u8; 4] = [0x28, 0xB5, 0x2F, 0xFD];
-
-/// Gzip magic number: 0x1F 0x8B.
-const GZIP_MAGIC: [u8; 2] = [0x1F, 0x8B];
-
-/// LZ4 frame magic number: 0x184D2204.
-const LZ4_MAGIC: [u8; 4] = [0x04, 0x22, 0x4D, 0x18];
+use crate::error::to_napi_error;
 
 /// Compression format detected from input data.
 #[napi(string_enum)]
@@ -44,12 +35,12 @@ pub enum CompressionFormat {
 #[napi]
 pub fn detect_format(data: Either<Buffer, Uint8Array>) -> CompressionFormat {
     let input = crate::as_bytes(&data);
-    match detect(input) {
-        Format::Zstd => CompressionFormat::Zstd,
-        Format::Gzip => CompressionFormat::Gzip,
-        Format::Brotli => CompressionFormat::Brotli,
-        Format::Lz4 => CompressionFormat::Lz4,
-        Format::Unknown => CompressionFormat::Unknown,
+    match comprs_core::detect::detect(input) {
+        comprs_core::detect::Format::Zstd => CompressionFormat::Zstd,
+        comprs_core::detect::Format::Gzip => CompressionFormat::Gzip,
+        comprs_core::detect::Format::Brotli => CompressionFormat::Brotli,
+        comprs_core::detect::Format::Lz4 => CompressionFormat::Lz4,
+        comprs_core::detect::Format::Unknown => CompressionFormat::Unknown,
     }
 }
 
@@ -63,69 +54,9 @@ pub fn detect_format(data: Either<Buffer, Uint8Array>) -> CompressionFormat {
 /// Raw deflate is not supported (no magic bytes to distinguish it).
 #[napi]
 pub fn decompress(data: Either<Buffer, Uint8Array>) -> Result<Buffer> {
-    let input = crate::as_bytes(&data);
-    match detect(input) {
-        Format::Zstd => crate::zstd_decompress(data),
-        Format::Gzip => crate::gzip_decompress(data),
-        Format::Brotli => crate::brotli_decompress(data),
-        Format::Lz4 => crate::lz4_decompress(data),
-        Format::Unknown => Err(ComprsError::InvalidArg(
-            "unable to detect compression format; use algorithm-specific functions (zstdDecompress, gzipDecompress, brotliDecompress, lz4Decompress) instead".to_string(),
-        )
-        .into()),
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum Format {
-    Zstd,
-    Gzip,
-    Brotli,
-    Lz4,
-    Unknown,
-}
-
-impl std::fmt::Display for Format {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Format::Zstd => write!(f, "zstd"),
-            Format::Gzip => write!(f, "gzip"),
-            Format::Brotli => write!(f, "brotli"),
-            Format::Lz4 => write!(f, "lz4"),
-            Format::Unknown => write!(f, "unknown"),
-        }
-    }
-}
-
-fn detect(data: &[u8]) -> Format {
-    if data.len() >= 4 && data[..4] == ZSTD_MAGIC {
-        return Format::Zstd;
-    }
-    if data.len() >= 2 && data[..2] == GZIP_MAGIC {
-        return Format::Gzip;
-    }
-    if data.len() >= 4 && data[..4] == LZ4_MAGIC {
-        return Format::Lz4;
-    }
-    // Brotli has no magic bytes. Attempt heuristic detection:
-    // A valid brotli stream starts with a window size byte where the
-    // lower nibble encodes WBITS. We try a quick decompression of the
-    // first few bytes to see if brotli accepts it.
-    if !data.is_empty() && is_likely_brotli(data) {
-        return Format::Brotli;
-    }
-    Format::Unknown
-}
-
-/// Heuristic check for brotli data.
-/// Attempts to decompress a small prefix to see if brotli accepts the header.
-fn is_likely_brotli(data: &[u8]) -> bool {
-    use std::io::Read;
-    let mut decompressor = brotli::Decompressor::new(data, 4096);
-    let mut buf = [0u8; 1];
-    // If brotli can parse the header and produce output (or cleanly reach EOF),
-    // it's likely a valid brotli stream.
-    decompressor.read_exact(&mut buf).is_ok()
+    comprs_core::detect::decompress(crate::as_bytes(&data))
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 pub struct DecompressTask {
@@ -138,58 +69,7 @@ impl Task for DecompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        match detect(&self.data) {
-            Format::Zstd => {
-                let capacity = match zstd::zstd_safe::get_frame_content_size(&self.data) {
-                    Ok(Some(size)) => (size as usize).min(crate::MAX_DECOMPRESSED_SIZE),
-                    _ => crate::MAX_DECOMPRESSED_SIZE,
-                };
-                let capacity = capacity.max(1024);
-                zstd::bulk::decompress(&self.data, capacity).map_err(|e| {
-                    ComprsError::Operation {
-                        context: "zstd decompress",
-                        source: e.into(),
-                    }
-                    .into()
-                })
-            }
-            Format::Gzip => {
-                let decoder = flate2::read::MultiGzDecoder::new(self.data.as_slice());
-                let init_cap = self.data.len().saturating_mul(4).min(crate::MAX_DECOMPRESSED_SIZE);
-                crate::decompress_with_limit(
-                    decoder,
-                    crate::MAX_DECOMPRESSED_SIZE,
-                    init_cap,
-                    "gzip decompress",
-                )
-            }
-            Format::Brotli => {
-                let decompressor =
-                    brotli::Decompressor::new(self.data.as_slice(), 4096);
-                let init_cap = self.data.len().saturating_mul(4).min(crate::MAX_DECOMPRESSED_SIZE);
-                crate::decompress_with_limit(
-                    decompressor,
-                    crate::MAX_DECOMPRESSED_SIZE,
-                    init_cap,
-                    "brotli decompress",
-                )
-            }
-            Format::Lz4 => {
-                let decoder = lz4_flex::frame::FrameDecoder::new(self.data.as_slice());
-                let init_cap =
-                    self.data.len().saturating_mul(4).min(crate::MAX_DECOMPRESSED_SIZE);
-                crate::decompress_with_limit(
-                    decoder,
-                    crate::MAX_DECOMPRESSED_SIZE,
-                    init_cap,
-                    "lz4 decompress",
-                )
-            }
-            Format::Unknown => Err(ComprsError::InvalidArg(
-                "unable to detect compression format; use algorithm-specific functions (zstdDecompress, gzipDecompress, brotliDecompress, lz4Decompress) instead".to_string(),
-            )
-            .into()),
-        }
+        comprs_core::detect::decompress(&self.data).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -209,59 +89,4 @@ impl Task for DecompressTask {
 pub fn decompress_async(data: Either<Buffer, Uint8Array>) -> AsyncTask<DecompressTask> {
     let input = crate::as_bytes(&data).to_vec();
     AsyncTask::new(DecompressTask { data: input })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Write;
-
-    #[test]
-    fn detect_zstd() {
-        let data = zstd::bulk::compress(b"test data for zstd", 3).unwrap();
-        assert_eq!(detect(&data), Format::Zstd);
-    }
-
-    #[test]
-    fn detect_gzip() {
-        let mut encoder = flate2::write::GzEncoder::new(Vec::new(), flate2::Compression::default());
-        encoder.write_all(b"test data for gzip").unwrap();
-        let data = encoder.finish().unwrap();
-        assert_eq!(detect(&data), Format::Gzip);
-    }
-
-    #[test]
-    fn detect_brotli() {
-        let mut output = Vec::new();
-        {
-            let mut compressor = brotli::CompressorWriter::new(&mut output, 4096, 6, 22);
-            compressor.write_all(b"test data for brotli").unwrap();
-        }
-        assert_eq!(detect(&output), Format::Brotli);
-    }
-
-    #[test]
-    fn detect_lz4() {
-        let mut compressed = Vec::new();
-        let mut encoder = lz4_flex::frame::FrameEncoder::new(&mut compressed);
-        encoder.write_all(b"test data for lz4").unwrap();
-        encoder.finish().unwrap();
-        assert_eq!(detect(&compressed), Format::Lz4);
-    }
-
-    #[test]
-    fn detect_unknown() {
-        let data = b"this is not compressed data at all";
-        assert_eq!(detect(data), Format::Unknown);
-    }
-
-    #[test]
-    fn detect_empty() {
-        assert_eq!(detect(b""), Format::Unknown);
-    }
-
-    #[test]
-    fn detect_too_short() {
-        assert_eq!(detect(&[0x00]), Format::Unknown);
-    }
 }
