+++ crates/core/src/gzip.rs
@@ -1,19 +1,10 @@
 //! Gzip and raw deflate compression and decompression.
 
-use std::io::Write;
-
-use flate2::Compression;
-use flate2::GzBuilder;
-use flate2::read::{DeflateDecoder, GzDecoder, MultiGzDecoder};
-use flate2::write::{DeflateEncoder, GzEncoder};
 use napi::Task;
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
 
-use crate::ComprsError;
-
-/// Default compression level for gzip/deflate (flate2 default = 6).
-const DEFAULT_LEVEL: u32 = 6;
+use crate::error::to_napi_error;
 
 /// Compress data using gzip.
 ///
@@ -21,28 +12,9 @@ const DEFAULT_LEVEL: u32 = 6;
 /// Level ranges from 0 (no compression) to 9 (best compression). Default is 6.
 #[napi]
 pub fn gzip_compress(data: Either<Buffer, Uint8Array>, level: Option<u32>) -> Result<Buffer> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if level > 9 {
-        return Err(ComprsError::InvalidArg(
-            "gzip compression level must be between 0 and 9".to_string(),
-        )
-        .into());
-    }
-    let input = crate::as_bytes(&data);
-
-    let mut encoder = GzEncoder::new(Vec::with_capacity(input.len()), Compression::new(level));
-    encoder.write_all(input).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "gzip compress",
-            source: e.into(),
-        })
-    })?;
-    encoder.finish().map(|v| v.into()).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "gzip compress",
-            source: e.into(),
-        })
-    })
+    comprs_core::gzip::compress(crate::as_bytes(&data), level)
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Options for customizing the gzip header during compression.
@@ -80,36 +52,13 @@ pub fn gzip_compress_with_header(
     header: GzipHeaderOptions,
     level: Option<u32>,
 ) -> Result<Buffer> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if level > 9 {
-        return Err(ComprsError::InvalidArg(
-            "gzip compression level must be between 0 and 9".to_string(),
-        )
-        .into());
-    }
-    let input = crate::as_bytes(&data);
-
-    let mut builder = GzBuilder::new();
-    if let Some(ref filename) = header.filename {
-        builder = builder.filename(filename.as_bytes());
-    }
-    if let Some(mtime) = header.mtime {
-        builder = builder.mtime(mtime);
-    }
-
-    let mut encoder = builder.write(Vec::with_capacity(input.len()), Compression::new(level));
-    encoder.write_all(input).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "gzip compress with header",
-            source: e.into(),
-        })
-    })?;
-    encoder.finish().map(|v| v.into()).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "gzip compress with header",
-            source: e.into(),
-        })
-    })
+    let core_header = comprs_core::gzip::GzipHeaderOptions {
+        filename: header.filename,
+        mtime: header.mtime,
+    };
+    comprs_core::gzip::compress_with_header(crate::as_bytes(&data), &core_header, level)
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Read gzip header metadata without fully decompressing the data.
@@ -117,24 +66,13 @@ pub fn gzip_compress_with_header(
 /// Returns the parsed header fields including filename, mtime, comment, OS, and extra data.
 #[napi]
 pub fn gzip_read_header(data: Either<Buffer, Uint8Array>) -> Result<GzipHeader> {
-    let input = crate::as_bytes(&data);
-    let decoder = GzDecoder::new(input);
-    let header = decoder.header().ok_or_else(|| {
-        napi::Error::from(ComprsError::InvalidArg(
-            "invalid gzip data: unable to parse header".to_string(),
-        ))
-    })?;
-
+    let h = comprs_core::gzip::read_header(crate::as_bytes(&data)).map_err(to_napi_error)?;
     Ok(GzipHeader {
-        filename: header
-            .filename()
-            .map(|b| String::from_utf8_lossy(b).into_owned()),
-        mtime: header.mtime(),
-        comment: header
-            .comment()
-            .map(|b| String::from_utf8_lossy(b).into_owned()),
-        os: header.operating_system(),
-        extra: header.extra().map(|b| b.to_vec().into()),
+        filename: h.filename,
+        mtime: h.mtime,
+        comment: h.comment,
+        os: h.os,
+        extra: h.extra.map(|b| b.into()),
     })
 }
 
@@ -145,7 +83,9 @@ pub fn gzip_read_header(data: Either<Buffer, Uint8Array>) -> Result<GzipHeader> 
 /// for larger data.
 #[napi]
 pub fn gzip_decompress(data: Either<Buffer, Uint8Array>) -> Result<Buffer> {
-    decompress_gzip_with_limit(crate::as_bytes(&data), crate::MAX_DECOMPRESSED_SIZE)
+    comprs_core::gzip::decompress(crate::as_bytes(&data))
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Decompress gzip-compressed data with explicit capacity.
@@ -157,14 +97,10 @@ pub fn gzip_decompress_with_capacity(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<Buffer> {
-    let cap = crate::validate_capacity(capacity)?;
-    decompress_gzip_with_limit(crate::as_bytes(&data), cap)
-}
-
-fn decompress_gzip_with_limit(input: &[u8], max_size: usize) -> Result<Buffer> {
-    let decoder = MultiGzDecoder::new(input);
-    let init_cap = (input.len().saturating_mul(4)).min(max_size);
-    crate::decompress_with_limit(decoder, max_size, init_cap, "gzip decompress").map(|v| v.into())
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
+    comprs_core::gzip::decompress_with_capacity(crate::as_bytes(&data), cap)
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Compress data using raw deflate (no gzip header/trailer).
@@ -173,28 +109,9 @@ fn decompress_gzip_with_limit(input: &[u8], max_size: usize) -> Result<Buffer> {
 /// Level ranges from 0 (no compression) to 9 (best compression). Default is 6.
 #[napi]
 pub fn deflate_compress(data: Either<Buffer, Uint8Array>, level: Option<u32>) -> Result<Buffer> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if level > 9 {
-        return Err(ComprsError::InvalidArg(
-            "deflate compression level must be between 0 and 9".to_string(),
-        )
-        .into());
-    }
-    let input = crate::as_bytes(&data);
-
-    let mut encoder = DeflateEncoder::new(Vec::with_capacity(input.len()), Compression::new(level));
-    encoder.write_all(input).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "deflate compress",
-            source: e.into(),
-        })
-    })?;
-    encoder.finish().map(|v| v.into()).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "deflate compress",
-            source: e.into(),
-        })
-    })
+    comprs_core::gzip::deflate_compress(crate::as_bytes(&data), level)
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Decompress raw deflate-compressed data.
@@ -204,7 +121,9 @@ pub fn deflate_compress(data: Either<Buffer, Uint8Array>, level: Option<u32>) ->
 /// for larger data.
 #[napi]
 pub fn deflate_decompress(data: Either<Buffer, Uint8Array>) -> Result<Buffer> {
-    decompress_deflate_with_limit(crate::as_bytes(&data), crate::MAX_DECOMPRESSED_SIZE)
+    comprs_core::gzip::deflate_decompress(crate::as_bytes(&data))
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Decompress raw deflate-compressed data with explicit capacity.
@@ -216,22 +135,17 @@ pub fn deflate_decompress_with_capacity(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<Buffer> {
-    let cap = crate::validate_capacity(capacity)?;
-    decompress_deflate_with_limit(crate::as_bytes(&data), cap)
-}
-
-fn decompress_deflate_with_limit(input: &[u8], max_size: usize) -> Result<Buffer> {
-    let decoder = DeflateDecoder::new(input);
-    let init_cap = (input.len().saturating_mul(4)).min(max_size);
-    crate::decompress_with_limit(decoder, max_size, init_cap, "deflate decompress")
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
+    comprs_core::gzip::deflate_decompress_with_capacity(crate::as_bytes(&data), cap)
         .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 // --- Async tasks ---
 
 pub struct GzipCompressTask {
     data: Vec<u8>,
-    level: u32,
+    level: Option<u32>,
 }
 
 #[napi]
@@ -240,22 +154,7 @@ impl Task for GzipCompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let mut encoder = GzEncoder::new(
-            Vec::with_capacity(self.data.len()),
-            Compression::new(self.level),
-        );
-        encoder.write_all(&self.data).map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "gzip compress",
-                source: e.into(),
-            })
-        })?;
-        encoder.finish().map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "gzip compress",
-                source: e.into(),
-            })
-        })
+        comprs_core::gzip::compress(&self.data, self.level).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -272,12 +171,12 @@ pub fn gzip_compress_async(
     data: Either<Buffer, Uint8Array>,
     level: Option<u32>,
 ) -> Result<AsyncTask<GzipCompressTask>> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if level > 9 {
-        return Err(ComprsError::InvalidArg(
+    // Validate level eagerly
+    let lvl = level.unwrap_or(comprs_core::gzip::DEFAULT_LEVEL);
+    if lvl > 9 {
+        return Err(to_napi_error(comprs_core::ComprsError::InvalidArg(
             "gzip compression level must be between 0 and 9".to_string(),
-        )
-        .into());
+        )));
     }
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(GzipCompressTask { data: input, level }))
@@ -293,14 +192,7 @@ impl Task for GzipDecompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let decoder = MultiGzDecoder::new(self.data.as_slice());
-        let init_cap = (self.data.len().saturating_mul(4)).min(crate::MAX_DECOMPRESSED_SIZE);
-        crate::decompress_with_limit(
-            decoder,
-            crate::MAX_DECOMPRESSED_SIZE,
-            init_cap,
-            "gzip decompress",
-        )
+        comprs_core::gzip::decompress(&self.data).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -321,7 +213,7 @@ pub fn gzip_decompress_async(data: Either<Buffer, Uint8Array>) -> AsyncTask<Gzip
 
 pub struct DeflateCompressTask {
     data: Vec<u8>,
-    level: u32,
+    level: Option<u32>,
 }
 
 #[napi]
@@ -330,22 +222,7 @@ impl Task for DeflateCompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let mut encoder = DeflateEncoder::new(
-            Vec::with_capacity(self.data.len()),
-            Compression::new(self.level),
-        );
-        encoder.write_all(&self.data).map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "deflate compress",
-                source: e.into(),
-            })
-        })?;
-        encoder.finish().map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "deflate compress",
-                source: e.into(),
-            })
-        })
+        comprs_core::gzip::deflate_compress(&self.data, self.level).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -362,12 +239,11 @@ pub fn deflate_compress_async(
     data: Either<Buffer, Uint8Array>,
     level: Option<u32>,
 ) -> Result<AsyncTask<DeflateCompressTask>> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if level > 9 {
-        return Err(ComprsError::InvalidArg(
+    let lvl = level.unwrap_or(comprs_core::gzip::DEFAULT_LEVEL);
+    if lvl > 9 {
+        return Err(to_napi_error(comprs_core::ComprsError::InvalidArg(
             "deflate compression level must be between 0 and 9".to_string(),
-        )
-        .into());
+        )));
     }
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(DeflateCompressTask { data: input, level }))
@@ -383,14 +259,7 @@ impl Task for DeflateDecompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let decoder = DeflateDecoder::new(self.data.as_slice());
-        let init_cap = (self.data.len().saturating_mul(4)).min(crate::MAX_DECOMPRESSED_SIZE);
-        crate::decompress_with_limit(
-            decoder,
-            crate::MAX_DECOMPRESSED_SIZE,
-            init_cap,
-            "deflate decompress",
-        )
+        comprs_core::gzip::deflate_decompress(&self.data).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -422,9 +291,8 @@ impl Task for GzipDecompressWithCapacityTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let decoder = MultiGzDecoder::new(self.data.as_slice());
-        let init_cap = (self.data.len().saturating_mul(4)).min(self.capacity);
-        crate::decompress_with_limit(decoder, self.capacity, init_cap, "gzip decompress")
+        comprs_core::gzip::decompress_with_capacity(&self.data, self.capacity)
+            .map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -441,7 +309,7 @@ pub fn gzip_decompress_with_capacity_async(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<AsyncTask<GzipDecompressWithCapacityTask>> {
-    let cap = crate::validate_capacity(capacity)?;
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(GzipDecompressWithCapacityTask {
         data: input,
@@ -460,9 +328,8 @@ impl Task for DeflateDecompressWithCapacityTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let decoder = DeflateDecoder::new(self.data.as_slice());
-        let init_cap = (self.data.len().saturating_mul(4)).min(self.capacity);
-        crate::decompress_with_limit(decoder, self.capacity, init_cap, "deflate decompress")
+        comprs_core::gzip::deflate_decompress_with_capacity(&self.data, self.capacity)
+            .map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -479,273 +346,10 @@ pub fn deflate_decompress_with_capacity_async(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<AsyncTask<DeflateDecompressWithCapacityTask>> {
-    let cap = crate::validate_capacity(capacity)?;
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(DeflateDecompressWithCapacityTask {
         data: input,
         capacity: cap,
     }))
-}
-
-#[cfg(test)]
-mod tests {
-    use std::io::Read;
-
-    use super::*;
-
-    #[test]
-    fn gzip_round_trip_basic() {
-        let original = b"Hello, comprs! This is a test of gzip compression.";
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let mut decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn gzip_round_trip_empty() {
-        let original = b"";
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let mut decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn gzip_round_trip_large() {
-        let original: Vec<u8> = (0..100_000).map(|i| (i % 256) as u8).collect();
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(&original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let mut decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(original, decompressed);
-        assert!(compressed.len() < original.len());
-    }
-
-    #[test]
-    fn gzip_compression_levels() {
-        let data = b"Repeating data for compression level testing. ".repeat(100);
-        let mut fast_enc = GzEncoder::new(Vec::new(), Compression::new(1));
-        fast_enc.write_all(&data).unwrap();
-        let fast = fast_enc.finish().unwrap();
-
-        let mut best_enc = GzEncoder::new(Vec::new(), Compression::new(9));
-        best_enc.write_all(&data).unwrap();
-        let best = best_enc.finish().unwrap();
-
-        assert!(best.len() <= fast.len());
-    }
-
-    #[test]
-    fn deflate_round_trip_basic() {
-        let original = b"Hello, comprs! This is a test of deflate compression.";
-        let mut encoder = DeflateEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let mut decoder = DeflateDecoder::new(compressed.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn gzip_compress_rejects_level_above_9() {
-        // Compression::new clamps internally, but our validation should
-        // catch out-of-range values before reaching flate2.
-        assert!(flate2::Compression::new(10).level() <= 9 || true);
-        // The napi wrapper validates; here we verify the boundary directly.
-        // Level 9 is the maximum valid value for flate2-based compression.
-    }
-
-    #[test]
-    fn deflate_compress_rejects_level_above_9() {
-        // Same boundary assertion as gzip — level 10+ should be rejected
-        // by the napi-exported functions. The pure-Rust tests confirm
-        // the valid range works correctly (see round-trip tests).
-    }
-
-    // Tests exercise the read_to_end + Take pattern used by decompress_with_limit.
-    // Direct calls to decompress_with_limit are avoided because napi::Error::Drop
-    // references napi symbols unavailable in standalone test binaries.
-
-    #[test]
-    fn gzip_take_read_to_end_within_limit() {
-        let original = b"Hello, size-limited gzip!";
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut output = Vec::new();
-        decoder.take(1025).read_to_end(&mut output).unwrap();
-        assert_eq!(original.as_slice(), output.as_slice());
-    }
-
-    #[test]
-    fn gzip_take_read_to_end_exceeds_limit() {
-        let limit: usize = 1024;
-        let original: Vec<u8> = vec![0u8; limit + 1];
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(&original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut output = Vec::new();
-        decoder
-            .take((limit as u64).saturating_add(1))
-            .read_to_end(&mut output)
-            .unwrap();
-        assert!(
-            output.len() > limit,
-            "Take should allow reading past the limit to detect overflow"
-        );
-    }
-
-    #[test]
-    fn gzip_take_read_to_end_exact_boundary() {
-        let original = vec![42u8; 1024];
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(&original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = MultiGzDecoder::new(compressed.as_slice());
-        let mut output = Vec::new();
-        decoder.take(1025).read_to_end(&mut output).unwrap();
-        assert_eq!(original, output);
-    }
-
-    #[test]
-    fn deflate_take_read_to_end_within_limit() {
-        let original = b"Hello, size-limited deflate!";
-        let mut encoder = DeflateEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = DeflateDecoder::new(compressed.as_slice());
-        let mut output = Vec::new();
-        decoder.take(1025).read_to_end(&mut output).unwrap();
-        assert_eq!(original.as_slice(), output.as_slice());
-    }
-
-    #[test]
-    fn deflate_take_read_to_end_exceeds_limit() {
-        let limit: usize = 1024;
-        let original: Vec<u8> = vec![0u8; limit + 1];
-        let mut encoder = DeflateEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(&original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = DeflateDecoder::new(compressed.as_slice());
-        let mut output = Vec::new();
-        decoder
-            .take((limit as u64).saturating_add(1))
-            .read_to_end(&mut output)
-            .unwrap();
-        assert!(
-            output.len() > limit,
-            "Take should allow reading past the limit to detect overflow"
-        );
-    }
-
-    #[test]
-    fn deflate_round_trip_empty() {
-        let original = b"";
-        let mut encoder = DeflateEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let mut decoder = DeflateDecoder::new(compressed.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn gzip_concatenated_round_trip() {
-        let part_a = b"Hello";
-        let part_b = b" World";
-
-        let mut enc_a = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        enc_a.write_all(part_a).unwrap();
-        let compressed_a = enc_a.finish().unwrap();
-
-        let mut enc_b = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        enc_b.write_all(part_b).unwrap();
-        let compressed_b = enc_b.finish().unwrap();
-
-        let mut concatenated = compressed_a;
-        concatenated.extend_from_slice(&compressed_b);
-
-        let mut decoder = MultiGzDecoder::new(concatenated.as_slice());
-        let mut decompressed = Vec::new();
-        decoder.read_to_end(&mut decompressed).unwrap();
-        assert_eq!(b"Hello World", decompressed.as_slice());
-    }
-
-    #[test]
-    fn gzip_compress_with_header_filename() {
-        let original = b"Hello, gzip header!";
-        let mut encoder = GzBuilder::new()
-            .filename("hello.txt")
-            .write(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = GzDecoder::new(compressed.as_slice());
-        let header = decoder.header().expect("header should be present");
-        assert_eq!(
-            header
-                .filename()
-                .map(|b| String::from_utf8_lossy(b).into_owned()),
-            Some("hello.txt".to_string())
-        );
-    }
-
-    #[test]
-    fn gzip_compress_with_header_mtime() {
-        let original = b"Mtime test";
-        let mtime_val: u32 = 1_700_000_000;
-        let mut encoder = GzBuilder::new()
-            .mtime(mtime_val)
-            .write(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = GzDecoder::new(compressed.as_slice());
-        let header = decoder.header().expect("header should be present");
-        assert_eq!(header.mtime(), mtime_val);
-    }
-
-    #[test]
-    fn gzip_read_header_default() {
-        let original = b"default header";
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::new(DEFAULT_LEVEL));
-        encoder.write_all(original).unwrap();
-        let compressed = encoder.finish().unwrap();
-
-        let decoder = GzDecoder::new(compressed.as_slice());
-        let header = decoder.header().expect("header should be present");
-        assert_eq!(header.mtime(), 0);
-        assert!(header.filename().is_none());
-        assert!(header.comment().is_none());
-    }
-
-    #[test]
-    fn gzip_read_header_invalid_data() {
-        let invalid = b"not gzip data";
-        let decoder = GzDecoder::new(invalid.as_slice());
-        assert!(decoder.header().is_none());
-    }
 }
