+++ crates/core/src/zstd.rs
@@ -4,10 +4,7 @@ use napi::Task;
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
 
-use crate::ComprsError;
-
-/// Default compression level for zstd (same as the C library default).
-const DEFAULT_LEVEL: i32 = 3;
+use crate::error::to_napi_error;
 
 /// Compress data using Zstandard.
 ///
@@ -17,29 +14,14 @@ const DEFAULT_LEVEL: i32 = 3;
 /// ratio for speed. Level 0 is equivalent to the default level (3).
 #[napi]
 pub fn zstd_compress(data: Either<Buffer, Uint8Array>, level: Option<i32>) -> Result<Buffer> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if !(-131072..=22).contains(&level) {
-        return Err(ComprsError::InvalidArg(
-            "zstd compression level must be between -131072 and 22".to_string(),
-        )
-        .into());
-    }
-    let input = crate::as_bytes(&data);
-
-    zstd::bulk::compress(input, level)
+    comprs_core::zstd::compress(crate::as_bytes(&data), level)
         .map(|v| v.into())
-        .map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd compress",
-                source: e.into(),
-            }
-            .into()
-        })
+        .map_err(to_napi_error)
 }
 
 pub struct ZstdCompressTask {
     data: Vec<u8>,
-    level: i32,
+    level: Option<i32>,
 }
 
 #[napi]
@@ -48,13 +30,7 @@ impl Task for ZstdCompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        zstd::bulk::compress(&self.data, self.level).map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd compress",
-                source: e.into(),
-            }
-            .into()
-        })
+        comprs_core::zstd::compress(&self.data, self.level).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -73,12 +49,12 @@ pub fn zstd_compress_async(
     data: Either<Buffer, Uint8Array>,
     level: Option<i32>,
 ) -> Result<AsyncTask<ZstdCompressTask>> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if !(-131072..=22).contains(&level) {
-        return Err(ComprsError::InvalidArg(
+    // Validate level eagerly
+    let lvl = level.unwrap_or(comprs_core::zstd::DEFAULT_LEVEL);
+    if !(-131072..=22).contains(&lvl) {
+        return Err(to_napi_error(comprs_core::ComprsError::InvalidArg(
             "zstd compression level must be between -131072 and 22".to_string(),
-        )
-        .into());
+        )));
     }
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(ZstdCompressTask { data: input, level }))
@@ -94,19 +70,7 @@ impl Task for ZstdDecompressTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let capacity = match zstd::zstd_safe::get_frame_content_size(&self.data) {
-            Ok(Some(size)) => (size as usize).min(crate::MAX_DECOMPRESSED_SIZE),
-            _ => crate::MAX_DECOMPRESSED_SIZE,
-        };
-        let capacity = capacity.max(1024);
-
-        zstd::bulk::decompress(&self.data, capacity).map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd decompress",
-                source: e.into(),
-            }
-            .into()
-        })
+        comprs_core::zstd::decompress(&self.data).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -132,26 +96,9 @@ pub fn zstd_decompress_async(data: Either<Buffer, Uint8Array>) -> AsyncTask<Zstd
 /// for larger data.
 #[napi]
 pub fn zstd_decompress(data: Either<Buffer, Uint8Array>) -> Result<Buffer> {
-    let input = crate::as_bytes(&data);
-
-    // Try to read the frame content size from the header
-    let capacity = match zstd::zstd_safe::get_frame_content_size(input) {
-        Ok(Some(size)) => (size as usize).min(crate::MAX_DECOMPRESSED_SIZE),
-        _ => crate::MAX_DECOMPRESSED_SIZE,
-    };
-
-    // Use at least 1KB initial capacity
-    let capacity = capacity.max(1024);
-
-    zstd::bulk::decompress(input, capacity)
+    comprs_core::zstd::decompress(crate::as_bytes(&data))
         .map(|v| v.into())
-        .map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd decompress",
-                source: e.into(),
-            }
-            .into()
-        })
+        .map_err(to_napi_error)
 }
 
 /// Decompress Zstandard-compressed data with explicit capacity.
@@ -163,22 +110,11 @@ pub fn zstd_decompress_with_capacity(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<Buffer> {
-    let cap = crate::validate_capacity(capacity)?;
-    let input = crate::as_bytes(&data);
-
-    zstd::bulk::decompress(input, cap)
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
+    comprs_core::zstd::decompress_with_capacity(crate::as_bytes(&data), cap)
         .map(|v| v.into())
-        .map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd decompress",
-                source: e.into(),
-            }
-            .into()
-        })
+        .map_err(to_napi_error)
 }
-
-/// Default maximum dictionary size (110 KB, zstd default).
-const DEFAULT_MAX_DICT_SIZE: usize = 110 * 1024;
 
 /// Train a zstd dictionary from sample data.
 ///
@@ -192,24 +128,18 @@ pub fn zstd_train_dictionary(
     max_dict_size: Option<f64>,
 ) -> Result<Buffer> {
     let max_size = max_dict_size
-        .map(crate::validate_capacity)
+        .map(|s| comprs_core::validate_capacity(s).map_err(to_napi_error))
         .transpose()?
-        .unwrap_or(DEFAULT_MAX_DICT_SIZE);
+        .unwrap_or(comprs_core::zstd::DEFAULT_MAX_DICT_SIZE);
 
     let sample_vecs: Vec<Vec<u8>> = samples
         .iter()
         .map(|s| crate::as_bytes(s).to_vec())
         .collect();
 
-    zstd::dict::from_samples(&sample_vecs, max_size)
+    comprs_core::zstd::train_dictionary(&sample_vecs, max_size)
         .map(|v| v.into())
-        .map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd dictionary training",
-                source: e.into(),
-            }
-            .into()
-        })
+        .map_err(to_napi_error)
 }
 
 /// Compress data using Zstandard with a pre-trained dictionary.
@@ -222,30 +152,9 @@ pub fn zstd_compress_with_dict(
     dict: Either<Buffer, Uint8Array>,
     level: Option<i32>,
 ) -> Result<Buffer> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if !(-131072..=22).contains(&level) {
-        return Err(ComprsError::InvalidArg(
-            "zstd compression level must be between -131072 and 22".to_string(),
-        )
-        .into());
-    }
-    let input = crate::as_bytes(&data);
-    let dict_bytes = crate::as_bytes(&dict);
-
-    let mut compressor =
-        zstd::bulk::Compressor::with_dictionary(level, dict_bytes).map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "zstd compressor init",
-                source: e.into(),
-            })
-        })?;
-
-    compressor.compress(input).map(|v| v.into()).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "zstd compress with dict",
-            source: e.into(),
-        })
-    })
+    comprs_core::zstd::compress_with_dict(crate::as_bytes(&data), crate::as_bytes(&dict), level)
+        .map(|v| v.into())
+        .map_err(to_napi_error)
 }
 
 /// Decompress Zstandard-compressed data that was compressed with a dictionary.
@@ -256,31 +165,9 @@ pub fn zstd_decompress_with_dict(
     data: Either<Buffer, Uint8Array>,
     dict: Either<Buffer, Uint8Array>,
 ) -> Result<Buffer> {
-    let input = crate::as_bytes(&data);
-    let dict_bytes = crate::as_bytes(&dict);
-
-    let capacity = match zstd::zstd_safe::get_frame_content_size(input) {
-        Ok(Some(size)) => (size as usize).min(crate::MAX_DECOMPRESSED_SIZE),
-        _ => crate::MAX_DECOMPRESSED_SIZE,
-    };
-    let capacity = capacity.max(1024);
-
-    let mut decompressor = zstd::bulk::Decompressor::with_dictionary(dict_bytes).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "zstd decompressor init",
-            source: e.into(),
-        })
-    })?;
-
-    decompressor
-        .decompress(input, capacity)
+    comprs_core::zstd::decompress_with_dict(crate::as_bytes(&data), crate::as_bytes(&dict))
         .map(|v| v.into())
-        .map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "zstd decompress with dict",
-                source: e.into(),
-            })
-        })
+        .map_err(to_napi_error)
 }
 
 /// Decompress Zstandard-compressed data that was compressed with a dictionary,
@@ -295,26 +182,14 @@ pub fn zstd_decompress_with_dict_with_capacity(
     dict: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<Buffer> {
-    let cap = crate::validate_capacity(capacity)?;
-    let input = crate::as_bytes(&data);
-    let dict_bytes = crate::as_bytes(&dict);
-
-    let mut decompressor = zstd::bulk::Decompressor::with_dictionary(dict_bytes).map_err(|e| {
-        napi::Error::from(ComprsError::Operation {
-            context: "zstd decompressor init",
-            source: e.into(),
-        })
-    })?;
-
-    decompressor
-        .decompress(input, cap)
-        .map(|v| v.into())
-        .map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "zstd decompress with dict",
-                source: e.into(),
-            })
-        })
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
+    comprs_core::zstd::decompress_with_dict_with_capacity(
+        crate::as_bytes(&data),
+        crate::as_bytes(&dict),
+        cap,
+    )
+    .map(|v| v.into())
+    .map_err(to_napi_error)
 }
 
 pub struct ZstdDecompressWithCapacityTask {
@@ -328,13 +203,8 @@ impl Task for ZstdDecompressWithCapacityTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        zstd::bulk::decompress(&self.data, self.capacity).map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd decompress",
-                source: e.into(),
-            }
-            .into()
-        })
+        comprs_core::zstd::decompress_with_capacity(&self.data, self.capacity)
+            .map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -351,7 +221,7 @@ pub fn zstd_decompress_with_capacity_async(
     data: Either<Buffer, Uint8Array>,
     capacity: f64,
 ) -> Result<AsyncTask<ZstdDecompressWithCapacityTask>> {
-    let cap = crate::validate_capacity(capacity)?;
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
     let input = crate::as_bytes(&data).to_vec();
     Ok(AsyncTask::new(ZstdDecompressWithCapacityTask {
         data: input,
@@ -362,7 +232,7 @@ pub fn zstd_decompress_with_capacity_async(
 pub struct ZstdCompressWithDictTask {
     data: Vec<u8>,
     dict: Vec<u8>,
-    level: i32,
+    level: Option<i32>,
 }
 
 #[napi]
@@ -371,19 +241,8 @@ impl Task for ZstdCompressWithDictTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let mut compressor = zstd::bulk::Compressor::with_dictionary(self.level, &self.dict)
-            .map_err(|e| {
-                napi::Error::from(ComprsError::Operation {
-                    context: "zstd compressor init",
-                    source: e.into(),
-                })
-            })?;
-        compressor.compress(&self.data).map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "zstd compress with dict",
-                source: e.into(),
-            })
-        })
+        comprs_core::zstd::compress_with_dict(&self.data, &self.dict, self.level)
+            .map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -401,12 +260,12 @@ pub fn zstd_compress_with_dict_async(
     dict: Either<Buffer, Uint8Array>,
     level: Option<i32>,
 ) -> Result<AsyncTask<ZstdCompressWithDictTask>> {
-    let level = level.unwrap_or(DEFAULT_LEVEL);
-    if !(-131072..=22).contains(&level) {
-        return Err(ComprsError::InvalidArg(
+    // Validate level eagerly
+    let lvl = level.unwrap_or(comprs_core::zstd::DEFAULT_LEVEL);
+    if !(-131072..=22).contains(&lvl) {
+        return Err(to_napi_error(comprs_core::ComprsError::InvalidArg(
             "zstd compression level must be between -131072 and 22".to_string(),
-        )
-        .into());
+        )));
     }
     let input = crate::as_bytes(&data).to_vec();
     let dict_bytes = crate::as_bytes(&dict).to_vec();
@@ -428,25 +287,7 @@ impl Task for ZstdDecompressWithDictTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let capacity = match zstd::zstd_safe::get_frame_content_size(&self.data) {
-            Ok(Some(size)) => (size as usize).min(crate::MAX_DECOMPRESSED_SIZE),
-            _ => crate::MAX_DECOMPRESSED_SIZE,
-        };
-        let capacity = capacity.max(1024);
-
-        let mut decompressor =
-            zstd::bulk::Decompressor::with_dictionary(&self.dict).map_err(|e| {
-                napi::Error::from(ComprsError::Operation {
-                    context: "zstd decompressor init",
-                    source: e.into(),
-                })
-            })?;
-        decompressor.decompress(&self.data, capacity).map_err(|e| {
-            napi::Error::from(ComprsError::Operation {
-                context: "zstd decompress with dict",
-                source: e.into(),
-            })
-        })
+        comprs_core::zstd::decompress_with_dict(&self.data, &self.dict).map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -481,13 +322,8 @@ impl Task for ZstdTrainDictionaryTask {
     type JsValue = Buffer;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        zstd::dict::from_samples(&self.samples, self.max_dict_size).map_err(|e| {
-            ComprsError::Operation {
-                context: "zstd dictionary training",
-                source: e.into(),
-            }
-            .into()
-        })
+        comprs_core::zstd::train_dictionary(&self.samples, self.max_dict_size)
+            .map_err(to_napi_error)
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
@@ -507,9 +343,9 @@ pub fn zstd_train_dictionary_async(
     max_dict_size: Option<f64>,
 ) -> Result<AsyncTask<ZstdTrainDictionaryTask>> {
     let max_size = max_dict_size
-        .map(crate::validate_capacity)
+        .map(|s| comprs_core::validate_capacity(s).map_err(to_napi_error))
         .transpose()?
-        .unwrap_or(DEFAULT_MAX_DICT_SIZE);
+        .unwrap_or(comprs_core::zstd::DEFAULT_MAX_DICT_SIZE);
     let sample_vecs: Vec<Vec<u8>> = samples
         .iter()
         .map(|s| crate::as_bytes(s).to_vec())
@@ -520,102 +356,45 @@ pub fn zstd_train_dictionary_async(
     }))
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+pub struct ZstdDecompressWithDictWithCapacityTask {
+    data: Vec<u8>,
+    dict: Vec<u8>,
+    capacity: usize,
+}
 
-    #[test]
-    fn round_trip_basic() {
-        let original = b"Hello, comprs! This is a test of zstd compression.";
-        let compressed = zstd::bulk::compress(original, DEFAULT_LEVEL).unwrap();
-        let decompressed = zstd::bulk::decompress(&compressed, original.len()).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
+#[napi]
+impl Task for ZstdDecompressWithDictWithCapacityTask {
+    type Output = Vec<u8>;
+    type JsValue = Buffer;
+
+    fn compute(&mut self) -> Result<Self::Output> {
+        comprs_core::zstd::decompress_with_dict_with_capacity(&self.data, &self.dict, self.capacity)
+            .map_err(to_napi_error)
     }
 
-    #[test]
-    fn round_trip_empty() {
-        let original = b"";
-        let compressed = zstd::bulk::compress(original, DEFAULT_LEVEL).unwrap();
-        let decompressed = zstd::bulk::decompress(&compressed, 1024).unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
+    fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
+        Ok(output.into())
     }
+}
 
-    #[test]
-    fn round_trip_large() {
-        let original: Vec<u8> = (0..100_000).map(|i| (i % 256) as u8).collect();
-        let compressed = zstd::bulk::compress(&original, DEFAULT_LEVEL).unwrap();
-        let decompressed = zstd::bulk::decompress(&compressed, original.len()).unwrap();
-        assert_eq!(original, decompressed);
-        // Compression should actually reduce size for repetitive data
-        assert!(compressed.len() < original.len());
-    }
-
-    #[test]
-    fn compression_levels() {
-        let data = b"Repeating data for compression level testing. ".repeat(100);
-        let fast = zstd::bulk::compress(&data, 1).unwrap();
-        let default = zstd::bulk::compress(&data, DEFAULT_LEVEL).unwrap();
-        let best = zstd::bulk::compress(&data, 19).unwrap();
-
-        // Higher levels should generally produce smaller output
-        assert!(best.len() <= default.len());
-        assert!(default.len() <= fast.len());
-    }
-
-    #[test]
-    fn level_zero_uses_default() {
-        let data = b"Level zero test data. ".repeat(50);
-        let with_zero = zstd::bulk::compress(&data, 0).unwrap();
-        let decompressed = zstd::bulk::decompress(&with_zero, data.len()).unwrap();
-        assert_eq!(data.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn negative_levels() {
-        let data = b"Negative level test data. ".repeat(50);
-        for level in [-1, -7, -50] {
-            let compressed = zstd::bulk::compress(&data, level).unwrap();
-            let decompressed = zstd::bulk::decompress(&compressed, data.len()).unwrap();
-            assert_eq!(data.as_slice(), decompressed.as_slice());
-        }
-    }
-
-    #[test]
-    fn level_22_max_standard() {
-        let data = b"Max level test data. ".repeat(50);
-        let compressed = zstd::bulk::compress(&data, 22).unwrap();
-        let decompressed = zstd::bulk::decompress(&compressed, data.len()).unwrap();
-        assert_eq!(data.as_slice(), decompressed.as_slice());
-    }
-
-    #[test]
-    fn dict_train_and_round_trip() {
-        // Generate sample data (JSON-like patterns)
-        let samples: Vec<Vec<u8>> = (0..100)
-            .map(|i| {
-                format!(
-                    r#"{{"id":{},"name":"user_{}","email":"user{}@example.com","active":{}}}"#,
-                    i,
-                    i,
-                    i,
-                    i % 2 == 0
-                )
-                .into_bytes()
-            })
-            .collect();
-
-        let dict = zstd::dict::from_samples(&samples, DEFAULT_MAX_DICT_SIZE).unwrap();
-        assert!(!dict.is_empty());
-
-        // Compress and decompress with dictionary
-        let original = br#"{"id":999,"name":"test_user","email":"test@example.com","active":true}"#;
-        let mut compressor = zstd::bulk::Compressor::with_dictionary(DEFAULT_LEVEL, &dict).unwrap();
-        let compressed = compressor.compress(original).unwrap();
-
-        let mut decompressor = zstd::bulk::Decompressor::with_dictionary(&dict).unwrap();
-        let decompressed = decompressor
-            .decompress(&compressed, original.len())
-            .unwrap();
-        assert_eq!(original.as_slice(), decompressed.as_slice());
-    }
+/// Asynchronously decompress Zstandard-compressed data that was compressed with a dictionary,
+/// with explicit capacity.
+///
+/// Use this when the decompressed size exceeds the default 256 MB limit.
+/// The `capacity` parameter specifies the maximum decompressed size in bytes.
+/// The same dictionary used for compression must be provided.
+#[napi]
+pub fn zstd_decompress_with_dict_with_capacity_async(
+    data: Either<Buffer, Uint8Array>,
+    dict: Either<Buffer, Uint8Array>,
+    capacity: f64,
+) -> Result<AsyncTask<ZstdDecompressWithDictWithCapacityTask>> {
+    let cap = comprs_core::validate_capacity(capacity).map_err(to_napi_error)?;
+    let input = crate::as_bytes(&data).to_vec();
+    let dict_bytes = crate::as_bytes(&dict).to_vec();
+    Ok(AsyncTask::new(ZstdDecompressWithDictWithCapacityTask {
+        data: input,
+        dict: dict_bytes,
+        capacity: cap,
+    }))
 }
